sql-tuner/requirements.txt
openai
python-dotenv
fastapi
uvicorn
requests
python-dotenv
